Move scrolling info to osu.Game
